Add machine learning prediction

+++ baseApp.py
@@ -29,14 +29,37 @@ class MainApp(MDApp):
     classes = ["brushing_teeth", "cutting_nails", "doing_laundry", "folding_clothes", "washing_dishes"]
     currentTaskString = None
     currentTaskCard = None
+    model = None
 
     def build(self):
         #Set colours for app
         self.theme_cls.theme_style = "Dark"
         self.theme_cls.primary_palette = "Red"
         
+        self.model = load_model("./model.h5")
+        
         #Set tasks items for drop down menu
         taskItems = [{"viewclass": "IconListItem", "text": "Brushing teeth", "on_release": lambda x="Brushing teeth": self.setItem(x)}, {"viewclass": "IconListItem", "text": "Cutting nails", "on_release": lambda x="Cutting nails": self.setItem(x)}, {"viewclass": "IconListItem", "text": "Doing laundry", "on_release": lambda x="Doing laundry": self.setItem(x)}, {"viewclass": "IconListItem", "text": "Folding clothes", "on_release": lambda x="Folding clothes": self.setItem(x)}, {"viewclass": "IconListItem", "text": "Washing dishes", "on_release": lambda x="Washing dishes": self.setItem(x)}]
+
+        #Load all widgets up front to stop app slowing down in app
+        #Make camera button in the left center of the card
+        self.cameraAnchor = AnchorLayout(anchor_x="left", anchor_y="center")
+        self.cameraButton = MDIconButton(icon="camera", on_release=self.setCurrentTask)
+
+        #Make a delete button in the right center of the card
+        self.deleteAnchor = AnchorLayout(anchor_x="right", anchor_y="center")
+        self.deleteButton = MDIconButton(icon="delete", on_release=self.deleteTask)
+
+        #Make task label to the left top of the card
+        self.labelAnchor = AnchorLayout(anchor_x="left", anchor_y="top")
+        self.cardLabel = MDLabel(height=50)
+
+        #Make no tasks message to home screen
+        self.noTasksAnchor =  AnchorLayout(anchor_x="center", anchor_y="center")
+        self.noTasksLabel = MDLabel(id="noTasksYet", text="You have no tasks yet. Click the plus icon in the top right to add a task :)", adaptive_size=True, bold=True)
+
+        #Get camera object
+        self.camera = self.root.ids.camera
         
         #Make drop down menu
         self.menu = MDDropdownMenu(
@@ -48,14 +71,18 @@ class MainApp(MDApp):
 
 
 
+
+
+    #############Screens#############
     def toTaskScreen(self):
         self.root.current = "addTaskScreen"
 
     def toHomeScreen(self):
         #Delete all images taken so far
         files = glob.glob('.\\takenImages\\*')
-        for f in files:
-            os.remove(f)
+        if (len(files) > 0):
+            for f in files:
+                os.remove(f)
         self.currentTaskCard = None
         self.currentTaskString = None
         self.root.current = "homeScreen"
@@ -74,6 +101,9 @@ class MainApp(MDApp):
 
 
 
+
+
+    #############Tasks#############
     #Sets the current item in the task drop down menu
     def setItem(self, textItem):
         self.root.ids.taskDropDown.set_item(textItem)
@@ -98,23 +128,16 @@ class MainApp(MDApp):
                 self.openSameTaskDialog()
                 return
 
-        #Add camera button in the left center of the card
-        cameraAnchor = AnchorLayout(anchor_x="left", anchor_y="center")
-        cameraButton = MDIconButton(id=cameraButtonId, icon="camera", on_release=self.setCurrentTask)
-        cameraAnchor.add_widget(cameraButton)
+        self.cameraButton.id = cameraButtonId
+        self.cameraAnchor.add_widget(self.cameraButton)
 
-        #Add a delete button in the right center of the card
-        deleteAnchor = AnchorLayout(anchor_x="right", anchor_y="center")
-        deleteButton = MDIconButton(icon="delete", on_release=self.deleteTask)
-        deleteAnchor.add_widget(deleteButton)
+        self.deleteAnchor.add_widget(self.deleteButton)
 
-        #Add task label to the left top of the card
-        labelAnchor = AnchorLayout(anchor_x="left", anchor_y="top")
-        cardLabel = MDLabel(text=taskString, height=50)
-        labelAnchor.add_widget(cardLabel)
+        self.cardLabel.text = taskString
+        self.labelAnchor.add_widget(self.cardLabel)
 
         #Add card with above widgets ^^^^
-        self.root.ids.homeScreenWidgetLayout.add_widget(MDCard(labelAnchor, cameraAnchor, deleteAnchor, id=cardId, padding=20))
+        self.root.ids.homeScreenWidgetLayout.add_widget(MDCard(self.labelAnchor, self.cameraAnchor, self.deleteAnchor, id=cardId, padding=20))
         self.toHomeScreen()
         
     def setCurrentTask(self, instance):
@@ -129,19 +152,17 @@ class MainApp(MDApp):
         self.root.ids.homeScreenWidgetLayout.remove_widget(instance.parent.parent)
         #Check if any widgets exist in homeScreenWidgetLayout (check if no tasks)
         if (len(self.root.ids.homeScreenWidgetLayout.children) == 0):
-            #Add no tasks message to home screen
-            anchor =  AnchorLayout(anchor_x="center", anchor_y="center")
-            label = MDLabel(id="noTasksYet", text="You have no tasks yet. Click the plus icon in the top right to add a task :)", adaptive_size=True, bold=True)
-            anchor.add_widget(label)
-            self.root.ids.homeScreenWidgetLayout.add_widget(anchor)
+            self.noTasksAnchor.add_widget(self.noTasksLabel)
+            self.root.ids.homeScreenWidgetLayout.add_widget(self.noTasksAnchor)
 
 
 
+
+
+    #############Dialogs#############
     def openSameTaskDialog(self):
-        #Check if dialog exists
-        if not self.sameTaskDialog:
-            #Make dialog with warning and ok button that closes dialog
-            self.sameTaskDialog = MDDialog(text="That task is already set. Please pick another one or return to the home screen.", buttons=[MDFlatButton(text="OK", on_release=self.closeSameTaskDialog)])
+        #Make dialog with warning and ok button that closes dialog
+        self.sameTaskDialog = MDDialog(text="That task is already set. Please pick another one or return to the home screen.", buttons=[MDFlatButton(text="OK", on_release=self.closeSameTaskDialog)])
         #Open dialog
         self.sameTaskDialog.open()
 
@@ -175,33 +196,33 @@ class MainApp(MDApp):
 
 
 
+
+    #############Camera#############
     def capture(self):
-        #Get camera object
-        camera = self.root.ids.camera
         #Get current time
         timeStr = time.strftime("%Y%m%d_%H%M%S")
         #Set file name with current time
         fileName = "IMG_{}.png".format(timeStr)
         #Save image to takenImages folder
-        camera.export_to_png(".\\takenImages\\{}".format(fileName))
+        self.camera.export_to_png(".\\takenImages\\{}".format(fileName))
         #Go to image screen
         self.toImageScreen(fileName)
 
     def acceptImage(self):
-        model = load_model("./model.h5")
         files = glob.glob('.\\takenImages\\*')
-        kerasImage = load_img(files[0])
+        kerasImage = load_img(files[0], target_size=(150, 150))
         x = img_to_array(kerasImage)
         x = preprocess_input(x)
         x = np.expand_dims(x, axis=0)
-        pred = model.predict(x)[0]
+        pred = self.model.predict(x)[0]
         maxPosition=np.argmax(pred) 
         predClass = self.classes[maxPosition]
-        if (predClass == self.currentTaskString.lower()):
+        lowerCurrentTaskString = self.currentTaskString.lower()
+        if (predClass == lowerCurrentTaskString):
             self.root.ids.homeScreenWidgetLayout.remove_widget(self.currentTaskCard)
             self.openDoneTaskDialog()
         else:
-            self.openFailedTaskDialog
+            self.openFailedTaskDialog()
 
     def deleteImage(self):
         #Delete image from folder 
@@ -209,7 +230,7 @@ class MainApp(MDApp):
         for f in files:
             os.remove(f)
         #Take back to camera screen
-        self.toCameraScreen(self.root.ids.deleteImageButton)
+        self.toCameraScreen()
 
 if __name__ == '__main__':
     app = MainApp()
